requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.115.0
greenlet==3.2.4
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
